+++ Changes.md
@@ -5,6 +5,7 @@
 HEAD
 ----------
 
+- Updated `Sidekiq::Process` API to provide capsule data. The `queues` and `weights` data will be removed from Redis in Sidekiq 8.1, as this data can now be found in the `capsules` element. [#6295]
 - Job iteration now exposes `current_object` for easy access within the `around_iteration` callback [#6774]
 - Fix JS race condition which could skip confirmation dialogs when Live Polling [#6768]
 - Fix edge case which could lose CurrentAttributes [#6767]